requirements.txt
djangorestframework-simplejwt==5.5.0
python-dotenv==1.1.0
django-cors-headers==4.7.0
pillow==11.2.1
phonenumbers==9.0.5
django-model-utils==5.0.0
datetimerange==2.3.1
gunicorn==23.0.0
django-leaflet==0.32.0
geopy==2.4.1
psycopg2-binary==2.9.10
djangorestframework-gis==1.1
mailjet-rest==1.4.0
django-channels==0.7.0
channels==3.0.5
channels-redis==3.4.1
django==5.2.1
drf-yasg==1.21.10
certifi==2025.4.26
celery>=4.4.7,<5.5.0
redis==3.5.3
drf-api-logger==1.1.19
func-timeout==4.3.5
faker==37.3.0
djangorestframework==3.15.2
django-filter==24.3
sshtunnel==0.4.0
json-log-formatter==0.5.1
django-prometheus==2.2.0
python-logstash==0.4.8
prometheus-client==0.15.0
twisted==22.10.0
django-silk==5.0.2
daphne==3.0.2
sentry-sdk>=2.29.1
drf-nested-routers==0.94.2
django-soft-delete==0.9.21
django-storages==1.14.6
boto3==1.38.19
pygments==2.19.1
drf-spectacular==0.28.0
drf-standardized-errors==0.14.1
fedapay==0.3.0
reportlab==4.4.0
qrcode==8.2
autopep8==2.3.2
flake8==7.2.0
trycourier==6.1.0
django-simple-history==3.8.0
